added test for user registration page
