+++ src/constants.tsx
@@ -1,0 +1,1 @@
+export const PHONE_NUMBER_REGEX = /^\d*$/; // Allow only digits (empty string is valid for clearing the input)
