+++ vite.config.ts
@@ -34,4 +34,5 @@ export default defineConfig({
       type: 'module',
     },
   })],
+
 })
